Add args and readme

+++ src/index.ts
@@ -1,33 +1,59 @@
-import WebSocket from "ws";
+import { readFile, writeFile } from "fs/promises";
+import yargs from 'yargs';
+
 import { text2text } from './text2text';
-import { writeFile } from "fs/promises";
 
-const softPromptTemplate = `Today is 12th April 2023, and Alice is sitting in the Bodleian Library, Oxford. Alice is a smart, honest, helpful, harmless assistant to Bob. Alice has instant access to an online encyclopaedia containing all the facts about the world. Alice never says common misconceptions, outdated information, lies, fiction, myths, jokes, or memes.
+const argv = yargs(process.argv.slice(2)).options({
+  baseUrl: {type: 'string', default: 'ws://192.168.0.29:7860/queue/join'},
+  prompt: {type: 'string', require: true},
+  softPromptPath: {type: 'string'},
+  promptReplacePath: {type: 'string'},
+ }).parseSync();
 
-Bob: __basePromptReplace__
-
-Alice: 
-`;
-const promptTemplate = 'Make a bullet point list of the typical subjects in paintings by famous painter __promptReplace__.';
-const promptReplaceArr = [
-  'Vincent van Gogh',
-  'Pablo Picasso',
-  'Jackson Pollock'
-];
 const outputArr: string[] = [];
 let count = 1;
 
 async function processPromptReplaceList() {
-  for (const promptReplace of promptReplaceArr) {
-    const basePrompt = promptTemplate.replace('__promptReplace__', promptReplace);
-    const prompt = softPromptTemplate.replace('__basePromptReplace__', basePrompt);
-    console.log(`⚙️ (${count}/${promptReplaceArr.length}) - Starting: "${basePrompt}"`)
-    const output = await text2text(prompt);
-    outputArr.push(output.replace(prompt, ''));
-    count += 1;
+  let softPromptTemplate = '';
+  let promptReplaceArr: string[] = [];
+
+  // Load soft prompt if passed in args
+  try {
+    if (argv.softPromptPath) {
+      softPromptTemplate = (await readFile(argv.softPromptPath)).toString();
+    }
+  } catch (e) {
+    console.error(`Error loading soft prompt from path: ${argv.softPromptPath}`);
+    console.error(e);
   }
-  console.log(`✅ ${promptReplaceArr.length} prompts generated successfully!!!`);
-  await writeFile('./output.txt', outputArr.join('\n\n\n'))
+
+  // Load prompt replace list if passed in args.
+  try {
+    if (argv.promptReplacePath) {
+      promptReplaceArr = (await readFile(argv.promptReplacePath)).toString().split(/\r?\n/);
+    }
+  } catch (e) {
+    console.error(`Error loading prompt replace from path: ${argv.promptReplacePath}`);
+    console.error(e);
+  }
+
+  if (promptReplaceArr.length) {
+    for (const promptReplace of promptReplaceArr) {
+      const basePrompt = argv.prompt.replace('__replace__', promptReplace);
+      const prompt = softPromptTemplate ? softPromptTemplate.replace('__prompt__', basePrompt) : basePrompt;
+      console.log(`⚙️ (${count}/${promptReplaceArr.length}) - Starting: "${basePrompt}"`)
+      const output = await text2text(prompt, argv.baseUrl);
+      outputArr.push(output.replace(prompt, ''));
+      count += 1;
+    }
+    await writeFile('./output.txt', outputArr.join('\n\n\n'));
+    console.log(`✅ ${promptReplaceArr.length} prompt outputs generated successfully!!!`);
+  } else {
+    const prompt = softPromptTemplate ? softPromptTemplate.replace('__basePromptReplace__', argv.prompt) : argv.prompt;
+    const output = await text2text(prompt, argv.baseUrl);
+    console.log(`✅ Prompt output generated successfully!!!`);
+    console.log(output);
+  }
 }
 
 processPromptReplaceList();
